Addition of Buttons and Layout Adjustments
Lots of changes to improve the UX design of the site.

The biggest change was the addition of the buttons at the bottom of each page to aid in navigation.

Before this addition, the user would have to scroll to the top of the screen after each page, close the page, then click the next page for more information.

Now adding the buttons at the bottom of each page creates a beautiful flow from one page to the next, creating a cohesive story.

Added "The Flights" to showcase an important element of the bar.

Fixed the address display on the home screen. As well as added a link to the Instagram page, and below that a "Get Here" button which opens directly to Google Maps location of the bar.

The biggest issue I'm having is the content itself. The photos Greg provided me are very high quality, but all of them are vertical, leading to difficulties in cropping the image appropariately. The original images of the template used a 3:1 ratio, while Greg's images are all 2:3. I had to make compromises by simply using a 16:9 crop for the images.